feat: add Triple Midscroll format with shared multi-midscroll base
  - Add TripleMidscroll format supporting 3 full screens (300dvh)
  - Refactor double-midscroll to use shared multi-midscroll-base
  - Create reusable multi-midscroll-base with CSS custom properties
  - Add localhost auto-allowlisting for development environments
  - Include playground examples for triple-midscroll testing
  - Maintain backward compatibility with existing double-midscroll API
