+++ playground/local-dev/index.html
@@ -555,6 +555,26 @@
     </advantage-wrapper>
     <div class="container mx-auto px-4 py-4">
         <section class="grid grid-cols-1 lg:gap-6 lg:grid-cols-2 lg:grid-rows-1">
+            <article class="py-4">
+                <h2 class="sans-serif text-xl font-bold mb-4">
+                    Triple Midscroll Demo Section
+                </h2>
+                <p>
+                    This section demonstrates the triple midscroll format,
+                    which provides three full screens of scrolling space
+                    for enhanced engagement and visual impact.
+                </p>
+            </article>
+        </section>
+    </div>
+    <advantage-wrapper>
+        <div slot="advantage-ad-slot">
+            <iframe src="./triple-midscroll/short.html" title="advantage-ad" style="border: 0; width: 0; height: 0"
+                scrolling="no"></iframe>
+        </div>
+    </advantage-wrapper>
+    <div class="container mx-auto px-4 py-4">
+        <section class="grid grid-cols-1 lg:gap-6 lg:grid-cols-2 lg:grid-rows-1">
 
             <article class="py-4">
                 <h2 class="sans-serif text-xl font-bold mb-4">
