+++ playground/local-dev/triple-midscroll/tall.html
@@ -1,0 +1,112 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Triple Midscroll Background</title>
+  <style>
+    html,
+    body {
+      margin: 0;
+      padding: 0;
+      height: 100%;
+      width: 100%;
+    }
+
+    .absolute {
+      position: absolute
+    }
+
+    .relative {
+      position: relative
+    }
+
+    .inset-0 {
+      top: 0;
+      bottom: 0;
+      left: 0;
+      right: 0
+    }
+
+    .h-2 {
+      height: .5rem
+    }
+
+    .w-full {
+      width: 100%
+    }
+
+    .h-full {
+      height: 100%
+    }
+
+    .grid {
+      display: grid
+    }
+
+    .grid-rows-3 {
+      grid-template-rows: 1fr 1fr 1fr
+    }
+
+    .waypoint-debug>div {
+      background: repeating-linear-gradient(135deg, #6b04fd 0, #6b04fd 10%, transparent 0, transparent 50%);
+      background-size: 3em 3em;
+      background-color: #d9ceff;
+      opacity: 1
+    }
+
+    .section {
+      width: 100%;
+      height: 33.333vh;
+    }
+
+    .section-top {
+      background-color: #6b04fd;
+    }
+
+    .section-middle {
+      background-color: #4317a6;
+    }
+
+    .section-bottom {
+      background-color: #2a0d6b;
+    }
+  </style>
+
+  <script type="module">
+    import { AdvantageCreativeMessenger } from "/src/advantage/messaging/creative-side";
+    import { AdvantageMessageAction, AdvantageFormatName } from "/src/types";
+    import logger from "/src/utils/logging";
+    // Enable debug mode dynamically
+    logger.enableDebugMode();
+
+    window.advantageMessenger = new AdvantageCreativeMessenger();
+    document.addEventListener("DOMContentLoaded", () => {
+      const waypoints = document.querySelectorAll('#waypoints > div');
+      window.advantageMessenger.setupWaypoints(waypoints);
+    });
+  </script>
+</head>
+
+<body style="transition: opacity 1s ease-in;" class="grid grid-rows-3">
+
+  <!-- Three sections stacked vertically -->
+  <div class="section section-top"></div>
+  <div class="section section-middle"></div>
+  <div class="section section-bottom"></div>
+
+  <!--
+    The waypoints are used to trigger animation.
+    The data-id attribute is used to identify the waypoint and the style attribute is used to set the position of the waypoint on the page.
+  -->
+  <div id="waypoints" class="absolute inset-0">
+    <div class="h-2 w-full absolute" data-id="start" style="top: 0%"></div>
+    <div class="h-2 w-full absolute" data-id="first-third" style="top: 33%"></div>
+    <div class="h-2 w-full absolute" data-id="midpoint" style="top: 50%"></div>
+    <div class="h-2 w-full absolute" data-id="second-third" style="top: 66%"></div>
+    <div class="h-2 w-full absolute" data-id="end" style="top: 99%"></div>
+  </div>
+</body>
+
+</html>
