+++ src/advantage/formats/index.ts
@@ -2,10 +2,12 @@ import { topscroll } from "./topscroll";
 import { midscroll } from "./midscroll";
 import { welcomePage } from "./welcomepage";
 import { doubleMidscroll } from "./double-midscroll";
+import { tripleMidscroll } from "./triple-midscroll";
 
 export const defaultFormats = [
     topscroll,
     midscroll,
     welcomePage,
-    doubleMidscroll
+    doubleMidscroll,
+    tripleMidscroll
 ];
